fix(auth): JWT authentication middleware
By default JWT token is not aware of tenants. Signing/verifying keys
are loaded once at server startup process. This means a token
generated for a user with user ID 1 in tenantA would allow access
to tenantB also with user ID 1. These 2 users are potentially different
persons, so it is a security risk.

Created a custom middleware that changes token_backend.signing_key
based on the tenant. The signing keys for tenants are stored in the
Client table in a newly added field "signing_key".